requirements.txt
# streamlit==1.31.1
# pandas>=2.2.0
# numpy>=1.24.0
# matplotlib>=3.8.0
# seaborn>=0.13.0
# plotly>=5.18.0
# statsmodels>=0.14.0
# scikit-learn>=1.4.0
# scipy>=1.12.0
# nltk>=3.8.1
# textblob>=0.17.1
# wordcloud>=1.9.3
# pydeck>=0.9.0
# altair>=5.2.0
# openpyxl>=3.1.0
# xlsxwriter>=3.1.0
# google-generativeai>=0.3.2
# networkx>=3.0


streamlit==1.31.1
pandas>=2.2.0
numpy>=1.24.0
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0
statsmodels>=0.14.0
scikit-learn>=1.4.0
scipy>=1.12.0
nltk>=3.8.1
textblob>=0.17.1
wordcloud>=1.9.3
pydeck>=0.9.0
altair>=5.2.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0
google-generativeai>=0.3.2
networkx>=3.0
supabase>=2.3.0
python-dotenv>=1.0.0